AdminServer/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
requests>=2.25
djangorestframework>=3.12
PyJWT>=1.7
python-jwt>=3.3